requirements-frozen.txt
bleach==3.1.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
docutils==0.15.2
gitdb2==2.0.6
GitPython==3.0.4
google-auth==1.6.3
idna==2.8
kubernetes==10.0.1
oauthlib==3.1.0
pkginfo==1.5.0.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
Pygments==2.4.2
python-dateutil==2.8.0
pywin32==225; sys_platform == 'win32'
PyYAML==5.1.2
readme-renderer==24.0
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
smmap2==2.0.5
tqdm==4.36.1
twine==2.0.0
urllib3==1.25.6
webencodings==0.5.1
websocket-client==0.56.0
xmlrunner==1.7.7
